Fixed 2 bugs related to reading/writing environment datagrams
Fixed a bug that caused the environment data to get lost for all pings after a partial environment datagram was read (up until this fix it was assumed EK80 would not write partial environment datagrams.) Partial environment datagrams are now handled correctly. This fix also required a change to parsing of environment datagrams. Also fixed a bug related to writing environment datagrams. pyEcholab doesn't store the original environment datagram times and the write_raw method was not creating suitable replacements resulting in environment datagrams that were close, but not in the correct place in the file. There were also possible issues in missing environment datagrams when rewriting files created with ping sequencing. These issues should be addressed. Added initial implementation of read_xyz in EK80 class to more closely align the xyz reading process with the bot process. Started to add a new simplified interface in the echosounder module to make basic tasks much easier and abstract much of the complexity of the underlying classes.