Only need one pulse compression silly
This set up a loop within a loop for the pulse compression, and np.convolve was running waaaay too many times, leading to the slow processing. With the current pyEcholab setup, the entire pdata is returned after match filter, then we just need to run the DFT for each window. Now ~1GB file done in 30 s.